fix vite config bugs
