+++ packages/client/src/modules/formBuilder/components/FieldBuilder/FieldBuilder.tsx
@@ -1,20 +1,21 @@
 import { FC, useEffect, useRef } from "react";
 import { useFormContext } from "react-hook-form";
+import styled from "@emotion/styled";
 import { InputAdornment } from "@mui/material";
 import TextField from "@mui/material/TextField";
 
 import { T_FormField } from "../../types";
 
-// const CssTextField = styled(TextField)({
-//   "& .MuiInputBase-root": {
-//     borderRadius: 10,
-//     backgroundColor: "white",
-//     padding: "3px 10px",
-//   },
-//   "& .MuiInputBase-input": {
-//     textAlign: "right",
-//   },
-// });
+const CssTextField = styled(TextField)({
+  "& .MuiInputBase-root": {
+    borderRadius: 10,
+    backgroundColor: "white",
+    padding: "3px 10px",
+  },
+  "& .MuiInputBase-input": {
+    textAlign: "right",
+  },
+});
 
 export const FieldBuilder: FC<T_FormField> = ({ label, name, type, disabled = false, defaultValue = "" }) => {
   const {
@@ -33,7 +34,7 @@ export const FieldBuilder: FC<T_FormField> = ({ label, name, type, disabled = fa
   const hasError = !!errors?.[name];
 
   return (
-    <TextField
+    <CssTextField
       variant="standard"
       InputProps={{
         startAdornment: <InputAdornment position="start">{label}</InputAdornment>,
