+++ packages/client/ssr.vite.config.ts
@@ -1,9 +1,5 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/naming-convention */
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
-import * as path from "path";
-
 import react from "@vitejs/plugin-react";
 import dotenv from "dotenv";
 import { defineConfig } from "vite";
@@ -14,13 +10,13 @@ dotenv.config();
 export default defineConfig({
   plugins: [react()],
   build: {
-    lib: {
-      entry: path.resolve(__dirname, "ssr.tsx"),
-      name: "Client",
-      formats: ["cjs"],
-    },
+    ssr: true,
     rollupOptions: {
+      input: {
+        ssr: "ssr.tsx",
+      },
       output: {
+        format: "cjs",
         dir: "dist-ssr",
       },
     },
