+++ packages/client/vite.config.ts
@@ -8,11 +8,11 @@ dotenv.config();
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  plugins: [react()],
   server: {
     port: Number(process.env.CLIENT_PORT) || 3000,
   },
   define: {
     __SERVER_PORT__: process.env.SERVER_PORT || 3001,
   },
-  plugins: [react()],
 });
